+++ lib/src/helpers/dbas_filesystem_platform_util_io.dart
@@ -1,3 +1,5 @@
 import 'dart:io';
 
-bool isTestEnvironment() => Platform.environment.containsKey('FLUTTER_TEST');
+final bool _isTest = Platform.environment.containsKey('FLUTTER_TEST');
+
+bool isTestEnvironment() => _isTest;